1.7.x: update to publish to central portal
Sonatype is phasing out OSSRH so update the publishing
to go to the new central portal.

Also updates scala, log4j, and spectator to recent versions
and fix various warnings.

+++ atlas-chart/src/main/scala/com/netflix/atlas/chart/graphics/Ticks.scala
@@ -395,7 +395,7 @@ object Ticks {
     if (logDistance <= n) {
       1
     } else {
-      math.ceil(logDistance / n).toInt
+      math.ceil(logDistance.toDouble / n).toInt
     }
   }
